requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.5.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
python-dotenv>=1.0.0

# Text processing and NLP
nltk>=3.8
sentence-transformers>=2.2.0
transformers>=4.21.0
torch>=1.12.0

# Vector database
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # Uncomment for GPU version

# HTTP client for LLM APIs (updated for compatibility)
httpx>=0.28.1

# Optional: LangChain for advanced text splitting
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

# Web interface (optional)
streamlit>=1.28.0
gradio>=4.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Keras for TensorFlow compatibility
keras>=3.5.0